+++ stories/badge.stories.ts
@@ -1,19 +1,29 @@
 import { html } from 'lit';
-import { Context, Story } from './story.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { defineComponents, IgcBadgeComponent } from '../src/index.js';
+import { Meta, StoryObj } from '@storybook/web-components';
+import { Context } from './story.js';
+
+defineComponents(IgcBadgeComponent);
 
 // region default
-const metadata = {
+const metadata: Meta<IgcBadgeComponent> = {
   title: 'Badge',
   component: 'igc-badge',
+  parameters: {
+    docs: {
+      description: {
+        component:
+          'The badge is a component indicating a status on a related item or an area\nwhere some active indication is required.',
+      },
+    },
+  },
   argTypes: {
     variant: {
       type: '"primary" | "info" | "success" | "warning" | "danger"',
       description: 'The type of badge.',
       options: ['primary', 'info', 'success', 'warning', 'danger'],
-      control: {
-        type: 'select',
-      },
+      control: { type: 'select' },
       defaultValue: 'primary',
     },
     outlined: {
@@ -26,23 +36,29 @@ const metadata = {
       type: '"rounded" | "square"',
       description: 'The shape of the badge.',
       options: ['rounded', 'square'],
-      control: {
-        type: 'inline-radio',
-      },
+      control: { type: 'inline-radio' },
       defaultValue: 'rounded',
     },
   },
+  args: { variant: 'primary', outlined: false, shape: 'rounded' },
 };
+
 export default metadata;
-interface ArgTypes {
+
+interface IgcBadgeArgs {
+  /** The type of badge. */
   variant: 'primary' | 'info' | 'success' | 'warning' | 'danger';
+  /** Sets whether to draw an outlined version of the badge. */
   outlined: boolean;
+  /** The shape of the badge. */
   shape: 'rounded' | 'square';
 }
+type Story = StoryObj<IgcBadgeArgs>;
+
 // endregion
 
-const Template: Story<ArgTypes, Context> = (
-  { outlined = false, shape, variant }: ArgTypes,
+const Template = (
+  { outlined = false, shape, variant }: IgcBadgeArgs,
   { globals: { direction } }: Context
 ) => {
   return html`
@@ -100,4 +116,4 @@ const Template: Story<ArgTypes, Context> = (
   `;
 };
 
-export const Basic = Template.bind({});
+export const Basic: Story = Template.bind({});