+++ stories/combo.stories.ts
@@ -1,15 +1,25 @@
 import { html } from 'lit';
-import { Context, Story } from './story.js';
+import { Context } from './story.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
-import { defineAllComponents, ComboItemTemplate } from '../src/index.js';
+import { ComboItemTemplate } from '../src/index.js';
 import { registerIconFromText } from '../src/components/icon/icon.registry';
+import { defineComponents, IgcComboComponent } from '../src/index.js';
+import { Meta, StoryObj } from '@storybook/web-components';
 
-defineAllComponents();
+defineComponents(IgcComboComponent);
+
+interface City {
+  id: string;
+  name: string;
+  zip: string;
+  country: string;
+}
 
 // region default
-const metadata = {
+const metadata: Meta<IgcComboComponent> = {
   title: 'Combo',
   component: 'igc-combo',
+  parameters: { docs: { description: {} } },
   argTypes: {
     name: {
       type: 'string',
@@ -78,9 +88,7 @@ const metadata = {
       type: '"ltr" | "rtl" | "auto"',
       description: 'The direction attribute of the control.',
       options: ['ltr', 'rtl', 'auto'],
-      control: {
-        type: 'inline-radio',
-      },
+      control: { type: 'inline-radio' },
       defaultValue: 'auto',
     },
     open: {
@@ -89,19 +97,13 @@ const metadata = {
       control: 'boolean',
       defaultValue: false,
     },
-    flip: {
-      type: 'boolean',
-      control: 'boolean',
-      defaultValue: true,
-    },
+    flip: { type: 'boolean', control: 'boolean', defaultValue: true },
     groupSorting: {
       type: '"asc" | "desc"',
       description:
         'Sorts the items in each group by ascending or descending order.',
       options: ['asc', 'desc'],
-      control: {
-        type: 'inline-radio',
-      },
+      control: { type: 'inline-radio' },
       defaultValue: 'asc',
     },
     caseSensitiveIcon: {
@@ -124,31 +126,70 @@ const metadata = {
       control: 'text',
     },
   },
+  args: {
+    disabled: false,
+    required: false,
+    invalid: false,
+    outlined: false,
+    singleSelect: false,
+    autofocusList: false,
+    placeholderSearch: 'Search',
+    dir: 'auto',
+    open: false,
+    flip: true,
+    groupSorting: 'asc',
+    caseSensitiveIcon: false,
+    disableFiltering: false,
+  },
 };
+
 export default metadata;
-interface ArgTypes {
+
+interface IgcComboArgs {
+  /** The name attribute of the control. */
   name: string;
+  /** The disabled attribute of the control. */
   disabled: boolean;
+  /** The required attribute of the control. */
   required: boolean;
+  /** The invalid attribute of the control. */
   invalid: boolean;
+  /** The outlined attribute of the control. */
   outlined: boolean;
+  /** Enables single selection mode and moves item filtering to the main input. */
   singleSelect: boolean;
+  /** The autofocus attribute of the control. */
   autofocus: boolean;
+  /** Focuses the list of options when the menu opens. */
   autofocusList: boolean;
+  /** The label attribute of the control. */
   label: string;
+  /** The placeholder attribute of the control. */
   placeholder: string;
+  /** The placeholder attribute of the search input. */
   placeholderSearch: string;
+  /** The direction attribute of the control. */
   dir: 'ltr' | 'rtl' | 'auto';
+  /** Sets the open state of the component. */
   open: boolean;
   flip: boolean;
+  /** Sorts the items in each group by ascending or descending order. */
   groupSorting: 'asc' | 'desc';
+  /** Enables the case sensitive search icon in the filtering input. */
   caseSensitiveIcon: boolean;
+  /** Disables the filtering of the list of options. */
   disableFiltering: boolean;
+  /**
+   * Returns the current selection as a list of commma separated values,
+   * represented by the display key, when provided.
+   */
   value: string;
 }
+type Story = StoryObj<IgcComboArgs>;
+
 // endregion
 
-(metadata as any).parameters = {
+Object.assign(metadata.parameters!, {
   actions: {
     handles: [
       'igcFocus',
@@ -160,14 +201,7 @@ interface ArgTypes {
       'igcChange',
     ],
   },
-};
-
-interface City {
-  id: string;
-  name: string;
-  zip: string;
-  country: string;
-}
+});
 
 const itemTemplate: ComboItemTemplate<City> = ({ item }) => {
   return html`
@@ -245,7 +279,7 @@ registerIconFromText(
   '<svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 0 24 24"><path d="M0 0h24v24H0z" fill="none"/><path d="M12 2C8.13 2 5 5.13 5 9c0 5.25 7 13 7 13s7-7.75 7-13c0-3.87-3.13-7-7-7zm0 9.5c-1.38 0-2.5-1.12-2.5-2.5s1.12-2.5 2.5-2.5 2.5 1.12 2.5 2.5-1.12 2.5-2.5 2.5z"/></svg>'
 );
 
-const Template: Story<ArgTypes, Context> = (
+const Template = (
   {
     name,
     disableFiltering,
@@ -262,7 +296,7 @@ const Template: Story<ArgTypes, Context> = (
     singleSelect = false,
     autofocusList,
     groupSorting = 'asc',
-  }: ArgTypes,
+  }: IgcComboComponent<City>,
   { globals: { direction } }: Context
 ) => html`
   <igc-combo
@@ -295,4 +329,4 @@ const Template: Story<ArgTypes, Context> = (
   </igc-combo>
 `;
 
-export const Basic = Template.bind({});
+export const Basic: Story = Template.bind({});