+++ stories/story.ts
@@ -1,15 +1,12 @@
-import { TemplateResult } from 'lit';
-
 export type Direction = 'ltr' | 'rtl' | 'auto';
 export type Variant = 'light' | 'dark';
-
-export interface Story<T, K> {
-  (args: T, context: K): TemplateResult;
-  args?: Partial<T>;
-  argTypes?: Record<string, unknown>;
-  context?: Partial<K>;
-}
+export type Size = 'attribute' | 'small' | 'medium' | 'large';
 
 export interface Context {
-  globals: { theme: string; direction: Direction; variant: Variant };
+  globals: {
+    theme: string;
+    direction: Direction;
+    variant: Variant;
+    size: Size;
+  };
 }