+++ stories/switch.stories.ts
@@ -1,11 +1,23 @@
 import { html } from 'lit';
-import { Context, Story } from './story.js';
+import { Context } from './story.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
+import { defineComponents, IgcSwitchComponent } from '../src/index.js';
+import { Meta, StoryObj } from '@storybook/web-components';
+
+defineComponents(IgcSwitchComponent);
 
 // region default
-const metadata = {
+const metadata: Meta<IgcSwitchComponent> = {
   title: 'Switch',
   component: 'igc-switch',
+  parameters: {
+    docs: {
+      description: {
+        component:
+          'Similar to a checkbox, a switch controls the state of a single setting on or off.',
+      },
+    },
+  },
   argTypes: {
     name: {
       type: 'string',
@@ -45,9 +57,7 @@ const metadata = {
       type: '"before" | "after"',
       description: 'The label position of the control.',
       options: ['before', 'after'],
-      control: {
-        type: 'inline-radio',
-      },
+      control: { type: 'inline-radio' },
       defaultValue: 'after',
     },
     ariaLabelledby: {
@@ -56,22 +66,41 @@ const metadata = {
       control: 'text',
     },
   },
+  args: {
+    disabled: false,
+    checked: false,
+    required: false,
+    invalid: false,
+    labelPosition: 'after',
+  },
 };
+
 export default metadata;
-interface ArgTypes {
+
+interface IgcSwitchArgs {
+  /** The name attribute of the control. */
   name: string;
+  /** The value attribute of the control. */
   value: string;
+  /** Disables the control. */
   disabled: boolean;
+  /** The checked state of the control. */
   checked: boolean;
+  /** Makes the control a required field. */
   required: boolean;
+  /** Controls the validity of the control. */
   invalid: boolean;
+  /** The label position of the control. */
   labelPosition: 'before' | 'after';
+  /** Sets the aria-labelledby attribute for the control. */
   ariaLabelledby: string;
 }
+type Story = StoryObj<IgcSwitchArgs>;
+
 // endregion
 
-const Template: Story<ArgTypes, Context> = (
-  { labelPosition, checked, disabled }: ArgTypes,
+const Template = (
+  { labelPosition, checked, disabled }: IgcSwitchArgs,
   { globals: { direction } }: Context
 ) => {
   return html`
@@ -86,4 +115,4 @@ const Template: Story<ArgTypes, Context> = (
   `;
 };
 
-export const Basic = Template.bind({});
+export const Basic: Story = Template.bind({});